move AbstractCUser and CUserManagers to separate files
